module6 xor: fix all-zero tree thresholds (float dtype) + Serial link
train_xor.py converted the Random Forest with emlearn's default int16 feature
quantization. Our XOR features live in [0,1], so the ~0.5 split thresholds
rounded to 0 and every tree collapsed to `features[i] < 0` — never true for
[0,1] inputs, so the generated model predicted a constant 0 (the "all weights
are 0" report). The corner check disagreed with sklearn ([0 0 0 0] vs [0 1 1 0]).

Convert with dtype="float": real thresholds are kept and the header emits
`const float *features`, which is exactly what xor-device/src/main.cpp already
passes. Corner check now matches sklearn ([0 1 1 0]).

Also fix a separate xor-device link failure (undefined reference to Serial):
the project pulls in no Adafruit libs, so the TinyUSB CDC that backs `Serial`
was never linked. Include <Adafruit_TinyUSB.h> (the BSP idiom, as in PDM.h).
xor-device now builds and links clean for wiscore_rak4631 (64 KB flash).

The generated xor_model.h stays uncommitted (generated at setup per platformio.ini).

+++ exercises/module6-emlearn/xor-train/train_xor.py
@@ -83,7 +83,14 @@ def main():
     # #include <eml_trees.h> from the generated header, so the emlearn include
     # dir (`python -c "import emlearn; print(emlearn.includedir)"`) must be
     # copied into the device project -> see README Part 1 / platformio.ini.
-    cmodel = emlearn.convert(model, method="inline")
+    #
+    # dtype='float' is REQUIRED here: our features live in [0, 1], and emlearn's
+    # default integer quantisation rounds the ~0.5 split thresholds to 0 — every
+    # tree collapses to `features[i] < 0`, which is never true for [0,1] inputs,
+    # so the converted model predicts a constant 0 (the corner check below then
+    # disagrees with sklearn). Float keeps the real thresholds; the device passes
+    # float features (xor-device/src/main.cpp), so the signature matches too.
+    cmodel = emlearn.convert(model, method="inline", dtype="float")
     cmodel.save(file="xor_model.h", name="xor_model")
     print("wrote xor_model.h  ->  cp xor_model.h ../xor-device/include/")
